requirements.txt
# Core dependencies
click>=8.0.0
pyyaml>=6.0
pytest>=7.0.0
pytest-cov>=4.0.0

# LinuxCNC HAL bindings (if available)
# hal-python>=0.1.0

# Optional: For better CLI experience
rich>=13.0.0
typer>=0.9.0

# Development tools
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0
